+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+Next Release
+------------
+* add input support for array-like objects, other objects via toString(), and decimal numbers stored as a string, #2
+
 0.2.1 / 2014-04-13
 ------------------
 * add base64 output, #5
-* npmignore the references, #4 
+* npmignore the references, #4
 
 
 0.2.0 / 2014-01-15